refactor: relocate import types to src/shared/types/import.ts
Move pure data types (ProgressUpdate, ImportResult, ImportOptions, etc.)
from src/main/import/types.ts and VcfPreviewResult + AnnotationType from
src/main/import/vcf/types.ts into a new src/shared/types/import.ts.
Replace originals with re-exports and update src/shared/types/api.ts to
import from shared instead of main, eliminating cross-boundary imports.
